api/requirements.txt
fastapi==0.109.2
python-multipart==0.0.9
sentence-transformers==2.5.1
scikit-learn==1.4.1.post1
python-docx==1.1.0
pdfplumber==0.11.0
google-generativeai==0.3.2
python-dotenv==1.0.1
mangum==0.17.0